move and rename itext
