+++ src/EventTypeDefs.ts
@@ -7,7 +7,7 @@ import type { Group } from './shapes/group.class';
 import type { TOriginX, TOriginY, TRadian } from './typedefs';
 import type { saveObjectTransform } from './util/misc/objectTransforms';
 import type { Canvas } from './canvas/canvas_events';
-import type { IText } from './shapes/itext.class';
+import type { IText } from './shapes/IText/IText';
 import type { StaticCanvas } from './canvas/static_canvas.class';
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface, @typescript-eslint/no-unused-vars
