+++ src/canvas/canvas.class.ts
@@ -31,7 +31,7 @@ import {
 } from '../typedefs';
 import { degreesToRadians } from '../util/misc/radiansDegreesConversion';
 import { getPointer, isTouchEvent } from '../util/dom_event';
-import type { IText } from '../shapes/itext.class';
+import type { IText } from '../shapes/IText/IText';
 import {
   cleanUpJsdomNode,
   makeElementUnselectable,
