+++ src/shapes/textbox.class.ts
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { TClassProperties } from '../typedefs';
-import { IText } from './itext.class';
+import { IText } from './IText/IText';
 import { textDefaultValues } from './text.class';
 import { classRegistry } from '../util/class_registry';
